pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nikcrm"
version = "0.1.0"
description = "Telegram employee registration bot and admin web interface"
authors = [{ name = "Nik CRM" }]
requires-python = ">=3.11"
dependencies = [
  # common
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "loguru>=0.7",
  "alembic>=1.13",
  "SQLAlchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  # bot
  "aiogram>=3.4",
  # web
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "jinja2>=3.1",
  "itsdangerous>=2.2",
  "python-jose[cryptography]>=3.3",
  "httpx>=0.27",
  # html helpers
  "python-multipart>=0.0.9"
]

[tool.setuptools]
packages = ["shared"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
